+++ spec/jwt/claims/audience_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe JWT::Claims::Audience do
     let(:scalar_aud) { 'ruby-jwt-aud' }
     let(:array_aud) { %w[ruby-jwt-aud test-aud ruby-ruby-ruby] }
 
-    subject(:verify!) { described_class.new(expected_audience: expected_audience).verify!(context: JWT::Claims::VerificationContext.new(payload: payload)) }
+    subject(:verify!) { described_class.new(expected_audience: expected_audience).verify!(context: SpecSupport::Token.new(payload: payload)) }
 
     context 'when the singular audience does not match' do
       let(:expected_audience) { 'no-match' }